+++ preload.js
@@ -77,6 +77,15 @@ contextBridge.exposeInMainWorld('api', {
   restoreDefaultSettings: () => ipcRenderer.invoke('restore-default-settings'),
   wipeLibraryData: () => ipcRenderer.invoke('wipe-library-data'),
 
+  // Device sync
+  syncGetInfo:    ()            => ipcRenderer.invoke('sync-get-info'),
+  syncCreate:     (pat)         => ipcRenderer.invoke('sync-create', pat),
+  syncConnect:    (gistId, pat) => ipcRenderer.invoke('sync-connect', gistId, pat),
+  syncPush:       ()            => ipcRenderer.invoke('sync-push'),
+  syncPull:       ()            => ipcRenderer.invoke('sync-pull'),
+  syncDisconnect: ()            => ipcRenderer.invoke('sync-disconnect'),
+  onSyncStatus:   (cb)          => ipcRenderer.on('sync-status', (_e, s) => cb(s)),
+
   // Window controls
   winMinimize: ()   => ipcRenderer.send('win-minimize'),
   winMaximize: ()   => ipcRenderer.send('win-maximize'),